BetterShader: don't concat shaders with different versions
The GLSL spec requires that #version is only ever used once in a shader.
This leads to problems when we concat multiple shaders (vertex, fragment,
etc.) to avoid compiling them seperately.

Some NVIDIA compilers allow multiple versions, other compilers will
refuse to compile the shaders.

To avoid these problems, we compile vertex, fragment, etc. shaders
seperately, also, we make a sanity check when simulating the
 #include directive (which isn't in the GLSL spec):
if a file is included, it must either have no version spec, or if
it does, it must be the same as the one of the file it was included from.
We could simply disallow #version in included files, but allowing it
isn't that much of a hassle.

Also, concatenating strings in Python is O(n^2), so the shaders
will now be concatenated using the recommended join(list) method.